+++ bin/epaa.py
@@ -425,6 +425,11 @@ def read_lig_ID_values(filename):
 
 def create_protein_column_value(pep):
     # retrieve Ensembl protein ID for given transcript IDs, if we want to provide additional protein ID types, adapt here
+
+    # we have to catch cases where no protein information is available, e.g. if there are issues on BioMart side
+    if transcriptProteinTable is None:
+        return ""
+
     all_proteins = [
         # split by : otherwise epytope generator suffix included
         transcriptProteinTable.query(f'transcript_id == "{transcript.transcript_id.split(":")[0]}"')["ensembl_id"]